libsparse: Use ssize_t/off_t for file length/offset
Because the image to be flashed might exceed 2G, even 4G, which will
be out of the range of int or uint type. So we need to use 64-bit int to
indicate the file length. With the new build option "-D_FILE_OFFSET_BITS=64",
ssize_t and off_t will be defined with int64_t.

+++ libsparse/backed_block.h
@@ -18,6 +18,7 @@
 #define _BACKED_BLOCK_H_
 
 #include <stdint.h>
+#include <unistd.h>
 
 struct backed_block_list;
 struct backed_block;
@@ -30,22 +31,22 @@ enum backed_block_type {
 };
 
 int backed_block_add_data(struct backed_block_list *bbl, void *data,
-		unsigned int len, unsigned int block);
+		ssize_t len, unsigned int block);
 int backed_block_add_fill(struct backed_block_list *bbl, unsigned int fill_val,
-		unsigned int len, unsigned int block);
+		ssize_t len, unsigned int block);
 int backed_block_add_file(struct backed_block_list *bbl, const char *filename,
-		int64_t offset, unsigned int len, unsigned int block);
+		off_t offset, ssize_t len, unsigned int block);
 int backed_block_add_fd(struct backed_block_list *bbl, int fd,
-		int64_t offset, unsigned int len, unsigned int block);
+		off_t offset, ssize_t len, unsigned int block);
 
 struct backed_block *backed_block_iter_new(struct backed_block_list *bbl);
 struct backed_block *backed_block_iter_next(struct backed_block *bb);
-unsigned int backed_block_len(struct backed_block *bb);
+ssize_t backed_block_len(struct backed_block *bb);
 unsigned int backed_block_block(struct backed_block *bb);
 void *backed_block_data(struct backed_block *bb);
 const char *backed_block_filename(struct backed_block *bb);
 int backed_block_fd(struct backed_block *bb);
-int64_t backed_block_file_offset(struct backed_block *bb);
+off_t backed_block_file_offset(struct backed_block *bb);
 uint32_t backed_block_fill_val(struct backed_block *bb);
 enum backed_block_type backed_block_type(struct backed_block *bb);
 int backed_block_split(struct backed_block_list *bbl, struct backed_block *bb,
